+++ scripts/wva-apk-icon-assemble-r.sh
@@ -1,11 +1,10 @@
 #!/bin/sh
 project_name="$1"
-icon_path="$2"
-assets_path="$3"
-template_param_path="$4"
+rsrc_android_path="$2"
+template_param_path="$3"
 
 if [ -z "$1" ]; then
-    echo "[usage] $0 project_name icon_path assets_path template_param_path"
+    echo "[usage] $0 project_name rsrc_android_path template_param_path"
     exit 0;
 fi
 mkdir -p target/apk-unsigned
@@ -30,8 +29,8 @@ fi
     find app/src/main/res -name "*.webp" -type f -delete
     #rm app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml
     rm app/src/main/res/mipmap-anydpi/*.xml
-    cp -r ../../../$icon_path/res app/src/main/
-    cp -r ../../../$assets_path/assets app/src/main/
+    cp -r ../../../$rsrc_android_path/res app/src/main/
+    cp -r ../../../$rsrc_android_path/assets app/src/main/
     ./gradlew assembleRelease
     # output: app/build/outputs/apk/release/app-release-unsigned.apk
     cp app/build/outputs/apk/release/app-release-unsigned.apk ../../../target/apk-unsigned/${project_name}-wva-app-release-unsigned.apk